requirements.txt
# Core image processing
pillow>=10.0.0

# AI & OpenAI
openai>=1.0.0

# FastAPI & server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart

# Data validation & configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Async & WebSocket
aiofiles
aiohttp
websockets

# ComfyUI integration
comfy_api_simplified

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4